requirements.txt
promql-parser>=0.7.0
pyyaml>=6.0
requests>=2.31
